updated permissions, user register

+++ openspy-web-backend/Models/Profile.cs
@@ -47,8 +47,8 @@ namespace CoreWeb.Models
         public int? Pic { get; set; }
         public int Userid { get; set; }
         public int Namespaceid { get; set; }
-        public sbyte Deleted { get; set; }
-        public sbyte Admin { get; set; }
+        public System.Int16 Deleted { get; set; }
+        public System.Int16 Admin { get; set; }
 
         public User User { get; set; }
